Make `check` parameters override `instance` parameters
This was what I originally had in mind. The only thing that I am iffy
about is overriding the `files` attribute. But, my thoughts are that it
would make sense but be confusing if we override everything except for
the `files` attribute. So, we just override all of the instance values
with the `check` values.